Wiring DB to Bot commands

+++ packages/infra/jukebotx_infra/db/models.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from uuid import UUID as PyUUID, uuid4
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func, BigInteger
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
@@ -37,10 +37,10 @@ class SubmissionModel(Base):
 
     id: Mapped[PyUUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid4)
     track_id: Mapped[PyUUID] = mapped_column(UUID(as_uuid=True), ForeignKey("tracks.id"), nullable=False)
-    guild_id: Mapped[int] = mapped_column(Integer, index=True, nullable=False)
-    channel_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    message_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    author_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    guild_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
+    channel_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    message_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    author_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     submitted_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
 
 
@@ -50,10 +50,14 @@ class QueueItemModel(Base):
     __tablename__ = "queue_items"
 
     id: Mapped[PyUUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid4)
-    guild_id: Mapped[int] = mapped_column(Integer, index=True, nullable=False)
+    guild_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     track_id: Mapped[PyUUID] = mapped_column(UUID(as_uuid=True), ForeignKey("tracks.id"), nullable=False)
-    requested_by: Mapped[int] = mapped_column(Integer, nullable=False)
+
+    # Discord user snowflake -> MUST be BigInteger
+    requested_by: Mapped[int] = mapped_column(BigInteger, nullable=False)
+
     status: Mapped[str] = mapped_column(String(32), index=True, nullable=False)
     position: Mapped[int] = mapped_column(Integer, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     updated_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
+
